Move WirelessModuleType into DataToTempCSV
