+++ DAS/das/utils/__init__.py
@@ -1,5 +1,4 @@
-from .TopicsEnum import WirelessModuleType
 from .DataToTempCSV import DataToTempCSV
 from .MockSensor import MockSensor
 
-__all__ = ["WirelessModuleType", "DataToTempCSV", "MockSensor"]
+__all__ = ["DataToTempCSV", "MockSensor"]
